preserve old tasks
