+++ src/rmplan/prompt.test.ts
@@ -1,9 +1,5 @@
 import { describe, expect, it } from 'bun:test';
-import {
-  generatePhaseStepsPrompt,
-  generateUpdatePrompt,
-  type PhaseGenerationContext,
-} from './prompt';
+import { generatePhaseStepsPrompt, type PhaseGenerationContext } from './prompt';
 
 describe('generatePhaseStepsPrompt', () => {
   it('should generate prompt with no previous phases', () => {
@@ -174,142 +170,5 @@ describe('generatePhaseStepsPrompt', () => {
     expect(prompt).toContain('Ensure all fields are properly populated');
     expect(prompt).toContain('Use proper YAML syntax with correct indentation');
     expect(prompt).toContain('Multi-line prompts should use the pipe (|) character');
-  });
-});
-
-describe('generateUpdatePrompt', () => {
-  it('should correctly embed planAsMarkdown and updateDescription in the prompt', () => {
-    const planAsMarkdown = `# My Test Plan
-
-## Goal
-To test the update prompt generation
-
-## Priority
-medium
-
-### Details
-This is a test plan with some details
-
----
-
-## Task: First Task
-**Description:** This is the first task
-**Files:**
-- src/test1.ts
-- src/test2.ts
-
-**Steps:**
-1.  **Prompt:**
-    \`\`\`
-    Do something first
-    \`\`\`
-2.  **Prompt:**
-    \`\`\`
-    Do something second
-    \`\`\``;
-
-    const updateDescription = 'Add a new task for error handling and update the priority to high';
-
-    const prompt = generateUpdatePrompt(planAsMarkdown, updateDescription);
-
-    // Check that the prompt contains the key sections
-    expect(prompt).toContain('# Plan Update Task');
-    expect(prompt).toContain(
-      'You are acting as a project manager tasked with updating an existing project plan'
-    );
-
-    // Check that the existing plan is embedded
-    expect(prompt).toContain('## Current Plan');
-    expect(prompt).toContain(planAsMarkdown);
-
-    // Check that the update description is embedded
-    expect(prompt).toContain('## Requested Update');
-    expect(prompt).toContain(updateDescription);
-
-    // Check instructions section
-    expect(prompt).toContain('## Instructions');
-    expect(prompt).toContain('Return the ENTIRE updated plan');
-    expect(prompt).toContain('For **Pending Tasks** only, you may:');
-    expect(prompt).toContain('Add new tasks');
-    expect(prompt).toContain('Remove existing pending tasks');
-    expect(prompt).toContain('Modify pending tasks');
-    expect(prompt).toContain('Preserve any unmodified parts');
-
-    // Check that it references the required output format
-    expect(prompt).toContain('## Required Output Format');
-    expect(prompt).toContain('Your response must follow the exact structure of the input plan');
-
-    // Check important notes
-    expect(prompt).toContain('## Important Notes');
-    expect(prompt).toContain('Output ONLY the updated plan in Markdown format');
-  });
-
-  it('should include instructions for preserving completed tasks', () => {
-    const planAsMarkdown = `# Test Plan
-
-## Goal
-Test goal
-
----
-
-# Completed Tasks
-*These tasks have been completed and should not be modified.*
-
-## Task: Completed Task [TASK-1] ✓
-**Description:** This task is done
-**Steps:** *(All completed)*
-1.  **Prompt:** ✓
-    \`\`\`
-    Completed step
-    \`\`\`
-
----
-
-# Pending Tasks
-*These tasks can be updated, modified, or removed as needed.*
-
-## Task: Pending Task [TASK-2]
-**Description:** This task is not done
-**Steps:**
-1.  **Prompt:**
-    \`\`\`
-    Pending step
-    \`\`\``;
-
-    const updateDescription = 'Add a new feature';
-
-    const prompt = generateUpdatePrompt(planAsMarkdown, updateDescription);
-
-    // Check for completed task preservation instructions
-    expect(prompt).toContain('CRITICAL: Preserve ALL completed tasks exactly as they appear');
-    expect(prompt).toContain('Completed tasks are marked with ✓');
-    expect(prompt).toContain('Do NOT modify, remove, or change any completed tasks');
-    expect(prompt).toContain('Keep all task IDs (e.g., [TASK-1], [TASK-2]) exactly as shown');
-
-    // Check for pending task instructions
-    expect(prompt).toContain('For **Pending Tasks** only, you may:');
-    expect(prompt).toContain('Add new tasks');
-    expect(prompt).toContain('Remove existing pending tasks');
-    expect(prompt).toContain('Modify pending tasks');
-
-    // Check for task numbering instructions
-    expect(prompt).toContain('Continue the task numbering sequence');
-    expect(prompt).toContain(
-      'if the last task is [TASK-5], new tasks should be [TASK-6], [TASK-7]'
-    );
-
-    // Check structure preservation
-    expect(prompt).toContain('Keep the "Completed Tasks" section if it exists');
-    expect(prompt).toContain('Keep the "Pending Tasks" section');
-    expect(prompt).toContain('Maintain the separation between completed and pending tasks');
-
-    // Check formatting requirements
-    expect(prompt).toContain('Task ID format [TASK-N]');
-    expect(prompt).toContain('Completed task markers (✓)');
-
-    // Check final warning
-    expect(prompt).toContain(
-      'NEVER modify completed tasks - they represent work that has already been done'
-    );
   });
 });
